week_7/day_4/TopMovies/app/src/main/java/example/codeclan/com/topmovies/Movie.java
package example.codeclan.com.topmovies;

public class Movie {

    private String title;
    private String genre;
    private int ranking;

    public Movie(String title, String genre, int ranking){
        this.title = title;
        this.genre = genre;
        this.ranking = ranking;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getGenre() {
        return genre;
    }

    public void setGenre(String genre) {
        this.genre = genre;
    }

    public int getRanking() {
        return ranking;
    }

    public void setRanking(int ranking) {
        this.ranking = ranking;
    }

    @Override
    public String toString(){
        return "Rank " + this.ranking + " movie is " + this.title + ". " + "It's genre is " + this.genre + ".";
    }
}
